+++ janitor/functions/conditional_join.py
@@ -42,7 +42,6 @@ def conditional_join(
     use_numba: bool = False,
     indicator: Optional[Union[bool, str]] = False,
     force: bool = False,
-    return_matching_indices: bool = False,
 ) -> pd.DataFrame:
     """The conditional_join function operates similarly to `pd.merge`,
     but supports efficient joins on inequality operators,
@@ -240,7 +239,6 @@ def conditional_join(
             - Numba support for equi join
         - 0.27.0
             - Added support for timedelta dtype.
-            - Added `return_matching_indices` parameter.
 
     Args:
         df: A pandas DataFrame.
@@ -280,8 +278,6 @@ def conditional_join(
             only appears in the right DataFrame, and `both` if the observation’s
             merge key is found in both DataFrames.
         force: If `True`, force the non-equi join conditions to execute before the equi join.
-        return_matching_indices: Returns only the matching indices for the left and right dataframes.
-            Applicable only for `how='inner'`.
 
 
     Returns:
@@ -300,7 +296,6 @@ def conditional_join(
         use_numba,
         indicator,
         force,
-        return_matching_indices,
     )
 
 
@@ -331,7 +326,7 @@ def _conditional_join_preliminary_checks(
     use_numba: bool,
     indicator: Union[bool, str],
     force: bool,
-    return_matching_indices,
+    return_matching_indices: bool = False,
 ) -> tuple:
     """
     Preliminary checks for conditional_join are conducted here.
@@ -385,8 +380,11 @@ def _conditional_join_preliminary_checks(
         check_column(right, [right_on])
         _check_operator(op)
 
-    if all(
-        (op == _JoinOperator.STRICTLY_EQUAL.value for *_, op in conditions)
+    if (
+        all(
+            (op == _JoinOperator.STRICTLY_EQUAL.value for *_, op in conditions)
+        )
+        and not return_matching_indices
     ):
         raise ValueError("Equality only joins are not supported.")
 
@@ -425,13 +423,6 @@ def _conditional_join_preliminary_checks(
 
     check("force", force, [bool])
 
-    check("return_matching_indices", return_matching_indices, [bool])
-
-    if return_matching_indices and (how != "inner"):
-        raise ValueError(
-            "return_matching_indices is applied only if how = 'inner'."
-        )
-
     return (
         df,
         right,
@@ -444,7 +435,6 @@ def _conditional_join_preliminary_checks(
         use_numba,
         indicator,
         force,
-        return_matching_indices,
     )
 
 
@@ -496,7 +486,7 @@ def _conditional_join_compute(
     use_numba: bool,
     indicator: Union[bool, str],
     force: bool,
-    return_matching_indices: bool,
+    return_matching_indices=False,
 ) -> pd.DataFrame:
     """
     This is where the actual computation
@@ -515,7 +505,6 @@ def _conditional_join_compute(
         use_numba,
         indicator,
         force,
-        return_matching_indices,
     ) = _conditional_join_preliminary_checks(
         df,
         right,
@@ -810,7 +799,7 @@ def _multiple_conditional_join_eq(
     right_on = [*right_on]
 
     rest = (
-        (df[left_on], right[right_on], op)
+        (left_on, right_on, op)
         for left_on, right_on, op in conditions
         if op != _JoinOperator.STRICTLY_EQUAL.value
     )
@@ -823,8 +812,14 @@ def _multiple_conditional_join_eq(
         sort=False,
     )._get_join_indexers()
 
-    if not left_index.size:
+    if not left_index.size or not rest:
         return None
+
+    rest = (
+        (df[left_on], right[right_on], op)
+        for left_on, right_on, op in conditions
+        if op != _JoinOperator.STRICTLY_EQUAL.value
+    )
 
     indices = _generate_indices(left_index, right_index, rest)
 
@@ -1304,4 +1299,51 @@ def _create_frame(
 
     return pd.concat(
         contents, axis=0, copy=False, sort=False, ignore_index=True
+    )
+
+
+def get_join_indices(
+    df: pd.DataFrame,
+    right: Union[pd.DataFrame, pd.Series],
+    conditions: list,
+    keep: Literal["first", "last", "all"] = "all",
+    use_numba: bool = False,
+    force: bool = False,
+) -> tuple[np.ndarray, np.ndarray]:
+    """Convenience function to return the matching indices from an inner join.
+
+    !!! info "New in version 0.27.0"
+
+    Args:
+        df: A pandas DataFrame.
+        right: Named Series or DataFrame to join to.
+        conditions: List of arguments of tuple(s) of the form
+            `(left_on, right_on, op)`, where `left_on` is the column
+            label from `df`, `right_on` is the column label from `right`,
+            while `op` is the operator.
+            The `col` class is also supported. The operator can be any of
+            `==`, `!=`, `<=`, `<`, `>=`, `>`. For multiple conditions,
+            the and(`&`) operator is used to combine the results
+            of the individual conditions.
+        use_numba: Use numba, if installed, to accelerate the computation.
+        keep: Choose whether to return the first match, last match or all matches.
+        force: If `True`, force the non-equi join conditions
+            to execute before the equi join.
+
+    Returns:
+        A tuple of indices for the rows in the dataframes that match.
+    """
+    return _conditional_join_compute(
+        df=df,
+        right=right,
+        conditions=conditions,
+        how="inner",
+        sort_by_appearance=False,
+        df_columns=None,
+        right_columns=None,
+        keep=keep,
+        use_numba=use_numba,
+        indicator=False,
+        force=force,
+        return_matching_indices=True,
     )