+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 -   [BUG] Fix logic for groupby in complete. Index support deprecated. Fix deprecation warning for fillna in `complete` PR #1289 @samukweku
 -   [ENH] `select` function now supports variable arguments - PR #1288 @samukweku
--   [ENH] `conditional_join` now supports timedelta dtype. Also, `return_matching_indices` parameter added, that allows only the matching indices to be returned. - PR #1297, Issue #1310
+-   [ENH] `conditional_join` now supports timedelta dtype. - PR #1297,
+-   [ENH]  `get_join_indices` function added - returns only join indices between two dataframes. Issue #1310
 
 ## [v0.26.0] - 2023-09-18